add API search with pagination

+++ src/main/java/com/lab/contactsmanagement/service/impl/ContactServiceImpl.java
@@ -1,19 +1,25 @@
 package com.lab.contactsmanagement.service.impl;
 
 import java.util.List;
-import java.util.Optional;
+import java.util.NoSuchElementException;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.lab.contactsmanagement.dto.ContactRequestDTO;
 import com.lab.contactsmanagement.entity.Contact;
+import com.lab.contactsmanagement.exceptions.ContactNotFoundException;
 import com.lab.contactsmanagement.repository.ContactRepository;
 import com.lab.contactsmanagement.service.ContactService;
 
 @Service
 public class ContactServiceImpl implements ContactService {
 
+	public static final int ROOT_CONTACT_PER_PAGE = 5;
+	
 	@Autowired
 	private ContactRepository contactRepository;
 
@@ -22,7 +28,8 @@ public class ContactServiceImpl implements ContactService {
 		Contact contactBuildData = new Contact();
 
 		if (contactId != null) {
-			contactBuildData = getContactById(contactId).get();
+			//Create Contact
+			contactBuildData = getContactById(contactId);
 			contactBuildData = Contact.builder()
 					.Id(contactId).name(contactRequest.getName())
 					.email(contactRequest.getEmail())
@@ -30,6 +37,7 @@ public class ContactServiceImpl implements ContactService {
 					.postalAddress(contactRequest.getPostalAddress())
 					.build();
 		} else {
+			//Update Contact
 			contactBuildData = Contact.builder().name(contactRequest.getName())
 					.email(contactRequest.getEmail())
 					.telephone(contactRequest.getTelephone())
@@ -43,7 +51,7 @@ public class ContactServiceImpl implements ContactService {
 
 	@Override
 	public List<Contact> getContacts() {
-		return contactRepository.findAll();
+		return null;
 	}
 
 	@Override
@@ -52,18 +60,35 @@ public class ContactServiceImpl implements ContactService {
 	}
 
 	@Override
-	public Optional<Contact> getContactById(Long id) {
-
-		Optional<Contact> optContact = contactRepository.findById(id);
-
-		if (optContact.isPresent())
-			return optContact;
-		throw new RuntimeException("No user by Id: " + id);
+	public Contact getContactById(Long id) throws ContactNotFoundException {
+		try {
+			return contactRepository.findById(id).get();
+		} catch (NoSuchElementException  ex) {
+			throw new ContactNotFoundException("Could not find any category with ID " + id);
+		}
 	}
 
 	@Override
 	public void deleteContact(Long id) {
+		Long countById = contactRepository.countById(id);
+		
+		if (countById == null || countById == 0) {
+			throw new RuntimeException("Could not find any contact with ID " + id);
+		}
+		
 		contactRepository.deleteById(id);
+	}
+
+	@Override
+	public Page<Contact> listByPage(int pageNum, String keyword) {
+
+		Pageable pageable = PageRequest.of(pageNum - 1, ROOT_CONTACT_PER_PAGE);
+
+		if (keyword != null) {
+			return contactRepository.findAll(keyword, pageable);
+		}
+
+		return contactRepository.findAll(pageable);		
 	}
 
 }
